Reorganize xmission directory
