+++ reVX/least_cost_xmission/__init__.py
@@ -2,6 +2,6 @@
 """
 Sub-packaged to compute the least cost path to tranmission
 """
-from .cost_creator import XmissionCostCreator
+from .costs.dry_cost_creator import XmissionCostCreator
 from .least_cost_paths import LeastCostPaths
 from .least_cost_xmission import LeastCostXmission
