+++ reVX/least_cost_xmission/costs/wet_cost_creator.py
@@ -10,7 +10,7 @@ import numpy.typing as npt
 from reVX.least_cost_xmission.config.constants import DEFAULT_DTYPE, \
     WET_COSTS_TIFF
 
-from reVX.least_cost_xmission.transmission_layer_io_handler import \
+from reVX.least_cost_xmission.layers.transmission_layer_io_handler import \
     TransLayerIoHandler
 
 logger = logging.getLogger(__name__)
